requirements.txt
amqp==2.1.4
argh==0.26.2
awscli==1.11.29
Babel==2.3.4
billiard==3.5.0.2
boto==2.45.0
botocore==1.4.86
cairocffi==0.8.0
CairoSVG==2.0.1
celery==4.0.2
cffi==1.9.1
colorama==0.3.7
cssselect==1.0.1
Django==1.10.4
django-crispy-forms==1.6.1
django-filter==1.0.1
django-geoposition==0.3.0
django-mathfilters==0.4.0
django-object-actions==0.9.0
django-storages==1.5.1
djangorestframework==3.5.4
docutils==0.13.1
flower==0.9.1
geopy==1.11.0
gunicorn==19.6.0
html5lib==0.999999999
jmespath==0.9.0
kombu==4.0.2
lxml==3.7.2
olefile==0.44
pathtools==0.1.2
Pillow==4.0.0
pisa==3.0.33
psycopg2==2.6.2
pyasn1==0.1.9
pycparser==2.17
pyPdf==1.13
PyPDF2==1.26.0
Pyphen==0.9.4
python-dateutil==2.6.0
python-http-client==2.2.1
pytz==2016.10
PyYAML==3.12
reportlab==3.3.0
rsa==3.4.2
s3transfer==0.1.9
sendgrid==3.6.3
sendgrid-django==4.0.4
six==1.10.0
tinycss==0.4
tornado==4.2
vine==1.1.3
WeasyPrint==0.34
webencodings==0.5
